requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.18.0
annotated-types==0.7.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==22.2.0
autograd==1.7.0
autograd-gamma==0.5.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blis==0.7.11
Bottleneck==1.4.0
brotlipy==0.7.0
cachetools==5.3.0
catalogue==2.0.10
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625807838443/work
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
charset-normalizer==3.1.0
click==8.1.7
cloudpathlib==0.20.0
cmake==3.26.0
comm==0.1.2
conda==4.10.3
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262148928/work
confection==0.1.5
contourpy==1.0.7
cryptography @ file:///tmp/build/80754af9/cryptography_1616769286105/work
cycler==0.11.0
cymem==2.0.10
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
ecos==2.0.14
einops==0.8.0
et-xmlfile==2.0.0
executing==1.2.0
fastai==2.7.18
fastcore==1.7.20
fastdownload==0.0.7
fastjsonschema==2.16.3
fastprogress==1.0.3
filelock==3.10.0
flatbuffers==24.3.25
focal-loss==0.0.7
fonttools==4.39.0
formulaic==1.0.2
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphlib-backport==1.1.0
grpcio==1.51.3
h5py==3.11.0
huggingface-hub==0.26.3
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
image-tabular==0.0.6
importlib-metadata==6.0.0
importlib-resources==5.12.0
interface-meta==1.3.0
invoke==2.2.0
ipykernel==6.21.3
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==8.0.3
jupyter-core==5.2.0
jupyter-events==0.6.3
jupyter-server==2.4.0
jupyter-server-fileid==0.8.0
jupyter-server-terminals==0.4.4
jupyter-server-ydoc==0.6.1
jupyter-ydoc==0.2.3
jupyterlab==3.6.1
jupyterlab-language-pack-zh-cn==3.6.post0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.20.0
jupyterlab-widgets==3.0.5
keras==2.13.1
kiwisolver==1.4.4
langcodes==3.4.1
language-data==1.3.0
libclang==18.1.1
lifelines==0.27.8
lightning-utilities==0.11.9
lit==15.0.7
marisa-trie==1.2.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.11
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.16.4
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.3
notebook-shim==0.2.2
numexpr==2.8.6
numpy==1.24.2
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
openpyxl==3.1.5
openslide-bin==4.0.0.6
openslide-python==1.4.1
opt-einsum==3.4.0
osqp==0.6.7.post3
packaging==23.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.1.1
preshed==3.0.9
prometheus-client==0.16.0
prompt-toolkit==3.0.38
propcache==0.2.0
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==2.10.3
pydantic-core==2.27.1
Pygments==2.14.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-box==7.2.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.1
qdldl==0.1.7.post4
regex==2024.11.6
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rsa==4.9
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
safetensors==0.4.5
scikit-learn==1.3.2
scikit-survival==0.22.2
scipy==1.10.1
Send2Trash==1.8.0
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1623709665295/work
smart-open==7.0.5
sniffio==1.3.0
soupsieve==2.4
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
supervisor==4.2.5
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
terminado==0.17.1
thinc==8.2.5
threadpoolctl==3.5.0
tinycss2==1.2.1
tokenizers==0.20.3
tomli==2.0.1
torch @ http://download.pytorch.org/whl/cu118/torch-2.0.0%2Bcu118-cp38-cp38-linux_x86_64.whl
torch-geometric==2.6.1
torchmetrics==1.5.2
torchsummary==1.5.1
torchvision @ http://download.pytorch.org/whl/cu118/torchvision-0.15.1%2Bcu118-cp38-cp38-linux_x86_64.whl
tornado==6.2
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
traitlets==5.9.0
transformers==4.46.3
triton==2.0.0
typer==0.15.1
typing-extensions==4.5.0
tzdata==2024.2
uri-template==1.2.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
wasabi==1.1.3
wcwidth==0.2.6
weasel==0.4.1
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.17.0
y-py==0.5.9
yarg==0.1.9
yarl==1.15.2
ypy-websocket==0.8.2
zipp==3.15.0
